refactor(logger, thread): rename log channel configuration parameters for clarity and introduce SpinLock class for thread synchronization

+++ base/sources/channel/channel_impl.h
@@ -11,7 +11,7 @@ namespace base
 namespace channel
 {
 
-constexpr uint8_t kMagic = 0x7F7F; // 魔数
+constexpr uint16_t kMagic = 0x7F7F; // 魔数
 
 enum EntryFlag : uint16_t
 {
